src/requirements.txt
gdown==4.4.0
numpy==1.20.1
Pillow==9.1.1
scikit_learn==1.1.1
torch==1.8.1
torchvision==0.9.1
wandb==0.12.16
